+++ CHANGES.md
@@ -8,8 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### New
+- Added "minimum uptime after recovery" feature: configurable
+  thresholds to keep recently-recovered BGP sessions in
+  WARNING/CRITICAL state until stable (inspired by cisco_bgp_peer plugin)
 
 ### Changed
+- When logging in via ssh wait for "assword: " and not just
+  for "assword:"
 
 ### Fixed